Refactor: Create namespace for deduplication, move files in
- Runner orchestrates deduplicating multiple patients based on a strategy
- Deduplicator merges 2 or more given patients
- Strategies is a list of ways to find duplicate patients
