+++ app/controllers/admin/deduplicate_patients_controller.rb
@@ -4,30 +4,15 @@ class Admin::DeduplicatePatientsController < AdminController
   def show
     authorize { current_admin.power_user? }
 
-    @patients = Patient.where(id: duplicate_patient_ids)
+    @patients = Patient.where(id: PatientDeduplication::Strategies.identifier_match(limit: 250).sample)
   end
 
   def merge
     authorize { current_admin.power_user? }
 
     duplicate_patients = Patient.where(id: params[:duplicate_patients])
-    deduped_patient = DeduplicatePatients.new(duplicate_patients).merge
+    deduped_patient = PatientDeduplication::Deduplicator.new(duplicate_patients).merge
 
     redirect_to admin_deduplicate_patients_path, notice: "Patient merged into #{deduped_patient.id}"
-  end
-
-  def duplicate_patient_ids
-    identifier_type = PatientBusinessIdentifier.identifier_types[:simple_bp_passport]
-
-    # This does an exact match based on case insensitive full name only.
-    PatientBusinessIdentifier
-      .select("identifier, array_agg(patient_id) as patient_ids")
-      .where.not(identifier: "")
-      .where(identifier_type: identifier_type)
-      .group("identifier")
-      .having("COUNT(distinct patient_id) > 1")
-      .limit(250)
-      .sample
-      .patient_ids
   end
 end
